addressed issues mentioned in the pull requests comments
- For the function compute_distance_between_images_with_angel() in src/plan_computation.py, changed the reduced formula using sin() and cos() back to its original form using tan() because that may be clearer to future readers.
- Moved fov() to src/camera_utils.py
- In main.ipynb, simplified the test code for non-nadir photos to print results for few selected angles.
- In main.ipynb, added text describing the implications of the various test cases such as increasing height and reducing overlap. Also combined the two test cases for reducing focal length because there is only one focal length instead of two independent focal lengths.
- Slight touch up with the diagrams.
